Bug 111953 - Clean up some b2g build warnings. r=shelly.lin,steven.lee,jack.hung,vincent.chang

+++ b2g/sidl/MobileManagerDelegate.cpp
@@ -91,7 +91,7 @@ MobileManagerDelegateService::ValidateCardId(int cardId) {
     return false;
   }
 
-  if (cardId >= numRadioInterfaces) {
+  if (uint32_t(cardId) >= numRadioInterfaces) {
     MOBILEDEBUG("Invalid cardId")
     return false;
   }
